Move RoomManager to sdk package
