+++ app/src/community/java/com/twilio/video/app/ui/room/CommunityRoomManagerModule.kt
@@ -7,6 +7,7 @@ import com.twilio.video.app.ApplicationScope
 import com.twilio.video.app.data.AuthServiceModule
 import com.twilio.video.app.data.DataModule
 import com.twilio.video.app.data.api.TokenService
+import com.twilio.video.app.sdk.RoomManager
 import dagger.Module
 import dagger.Provides
